+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - Remove `"settings"` parameter from `SequencerCompiler.__init__` to prevent confusion with `sequencer_cfg`. (!1068)
   - Add SimpleNumericalPulse to the Qblox pulse library. (!1074)
   - Fix `StitchedPulseBuilder` not taking its `t0` into account for VoltageOffsets. (!1089)
+  - Add stack_pulses compilation node to the Qblox backend. (!1084)
 - Compilation
   - Add internal logic to calculate pulse compensations. (!1070)
   - Refactor parts of the NCO/LO frequency determining logic, and remove the `parent` attribute from `InstrumentCompiler` and all subclasses. (!1069)